code clean up and dead code removal
