+++ src/GenFu.Web/Startup.cs
@@ -24,31 +24,20 @@ namespace GenFu.Web
                 s.IdleTimeout = TimeSpan.FromMinutes(1);
             });
 
-			//var b = new Microsoft.DotNet.ProjectModel.Compilation.LibraryExporter(new ProjectDescription("project", Directory.GetCurrentDirectory()), new Microsoft.DotNet.ProjectModel.Resolution.LibraryManager())
-
-			//var a = new Microsoft.VisualStudio.Web.CodeGeneration.DotNet.LibraryExporter(ProjectContext.Create(Directory.GetCurrentDirectory(), new NuGet.Frameworks.NuGetFramework(NuGetFramework.AnyFramework)), new ApplicationInfo("test", Directory.GetCurrentDirectory()));
-
-			services.AddScoped<ILibraryExporter, Microsoft.VisualStudio.Web.CodeGeneration.DotNet.LibraryExporter>(x => {
-				return new Microsoft.VisualStudio.Web.CodeGeneration.DotNet.LibraryExporter(ProjectContext.Create(Directory.GetCurrentDirectory(), new NuGet.Frameworks.NuGetFramework(NuGetFramework.AnyFramework)), new ApplicationInfo("test", Directory.GetCurrentDirectory()));
-			});
-
+			//add a Assembly Loader to services
 			services.AddScoped<AssemblyLoadContext, AssemblyLoadContext>(x => {
 				return AssemblyLoadContext.Default;
 			});
-
         }
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
         {
-
             if (env.IsDevelopment())
             {
                 app.UseBrowserLink();
                 app.UseDeveloperExceptionPage();
             }
 
-            // Add the platform handler to the request pipeline.
-           // app.UseIISPlatformHandler();
             app.UseStaticFiles();
             app.UseSession();
 
@@ -58,31 +47,21 @@ namespace GenFu.Web
                     name: "default",
                     template: "{controller=Home}/{action=Index}/{id?}");
             });
-
         }
     }
 
 	public class Program
 	{
-		public static IWebHostBuilder getBuilder(string[] args)
+		public static void Main(string[] args)
 		{
 			var host = new WebHostBuilder()
 				.UseContentRoot(Directory.GetCurrentDirectory())
 				.UseIISIntegration()
 				.UseKestrel()
 				.UseStartup<Startup>();
-			//.Build();
 
-			return host;
-		}
-
-		public static void Main(string[] args)
-		{
-			var host = getBuilder(args);
-
+			//launch the website!
 			host.Build().Run();
-
-			//todo make do keep alive and logging here
 		}
 	}
 }
